Extract assessment draft validation boundary
