+++ src/lib/quiz-markdown.ts
@@ -1,5 +1,5 @@
 import { validateAssessmentOptions } from '@/lib/assessments'
-import type { AssessmentDraftContent, AssessmentDraftQuestion } from '@/lib/server/assessment-drafts'
+import type { AssessmentDraftContent, AssessmentDraftQuestion } from '@/types'
 
 const UUID_RE =
   /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i
